versions: cloud-hypervisor: Bump to version 6d30fe05
The cloud-hypervisor commit `6d30fe05` introduced a fix on its API for
VFIO device hotplug (`VmAddDevice`), which is required for supporting
VFIO unplug through openAPI calls in kata.

+++ src/runtime/virtcontainers/pkg/cloud-hypervisor/client/model_memory_config.go
@@ -18,6 +18,6 @@ type MemoryConfig struct {
 	Shared bool `json:"shared,omitempty"`
 	Hugepages bool `json:"hugepages,omitempty"`
 	Balloon bool `json:"balloon,omitempty"`
-	BalloonSize int32 `json:"balloon_size,omitempty"`
+	BalloonSize int64 `json:"balloon_size,omitempty"`
 	Zones []MemoryZoneConfig `json:"zones,omitempty"`
 }
